Eidt File

+++ Models/Payment.cs
@@ -12,7 +12,7 @@ namespace GroupProject.Models
     {
         [Key]
         public int PaymentID { get; set; }
-        public double AmountPaid { get; set; }
+        public decimal AmountPaid { get; set; }
         public DateTime PaymentDate { get; set; }
         public string PaymentMethod { get; set; }
         public decimal RemainingBalance { get; set; }
